improvement: migrate guide book
