+++ apps/kg/modules/guide/guide-book/content/index.ts
@@ -1,0 +1,2 @@
+export * from './content';
+export * from './store';
