+++ apps/kg/modules/guide/guide-book/content/store.ts
@@ -1,9 +1,6 @@
-import React, { ReactElement } from 'react';
-import Book from '../assets/book.svg';
-import Image from 'next/image';
-import Link from 'next/link';
+import Book from '../../assets/book.svg';
 
-const dummyGuide = [
+export const dummyGuide = [
   {
     title: 'Panduan pendaftaran',
     img: Book,
@@ -50,27 +47,3 @@ const dummyGuide = [
     href: '/panduan/buku-panduan/panduan-pendaftaran',
   },
 ];
-
-export const ContentSection = (): ReactElement => {
-  return (
-    <div className="w-full px-6 md:px-8 lg:px-10">
-      <div className="title">
-        <h1 className="font-[700] text-[16px]">Buku panduan lainnya</h1>
-      </div>
-      <div className="mt-[24px] mb-[54px]">
-        {dummyGuide.map((item, index) => {
-          return (
-            <Link key={index} href={item.href}>
-              <div className="content bg-white dark:bg-[#232529]  w-full h-[54px] shadow-md rounded-[8px] flex justify-between items-center p-[24px] mt-[16px]">
-                <h1 className="font-[600] text-[16px] dark:text-white">
-                  {item.title}
-                </h1>
-                <Image src={item.img} alt={'book'} />
-              </div>
-            </Link>
-          );
-        })}
-      </div>
-    </div>
-  );
-};
