+++ apps/kg/modules/guide/guide-book/hero/index.ts
