+++ apps/kg/modules/guide/guide-book/module.tsx
@@ -1,9 +1,9 @@
 import { ReactElement } from 'react';
-import { ContentSection } from './content';
-import { HeroSection } from './hero';
+import { ContentSection } from './content/content';
+import { HeroSection } from './hero/hero';
 import { BaseLayout } from '../../common';
 import { BreadCrumb } from '@mknows-frontend-services/components/atoms';
-import { guideBookBreadCumbs } from './constant';
+import { guideBookBreadCumbs } from './const';
 
 export const GuideBookModule = (): ReactElement => {
   return (
