+++ apps/kg/modules/guide/landing/const.ts
