+++ apps/kg/modules/guide/landing/faq/index.ts
