+++ apps/kg/modules/guide/landing/guide/index.ts
