+++ apps/kg/modules/guide/landing/hero/hero.tsx
