+++ apps/kg/modules/guide/landing/hero/index.ts
@@ -1,0 +1,1 @@
+export * from './hero';
