+++ apps/kg/modules/guide/landing/index.ts
@@ -2,4 +2,4 @@ export * from './module';
 export * from './loading';
 export * from './faq';
 export * from './guide';
-export * from './landing';
+export * from './hero';
