+++ apps/kg/modules/guide/landing/loading.tsx
